+++ weblate/appsettings.py
@@ -114,6 +114,7 @@ CHECK_LIST = getvalue('CHECK_LIST', (
     'weblate.trans.checks.format.PythonBraceFormatCheck',
     'weblate.trans.checks.format.PHPFormatCheck',
     'weblate.trans.checks.format.CFormatCheck',
+    'weblate.trans.checks.format.JavascriptFormatCheck',
     'weblate.trans.checks.consistency.PluralsCheck',
     'weblate.trans.checks.consistency.ConsistencyCheck',
     'weblate.trans.checks.chars.NewlineCountingCheck',